# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API REST du Task Manager — gestion complète des tâches (CRUD)
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Statut possible d'une tâche
 */
export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  À_faire: 'À faire',
  En_cours: 'En cours',
  Terminée: 'Terminée',
} as const;

export interface Task {
  /** Identifiant unique */
  id: string;
  /**
     * Titre de la tâche
     * @maxLength 200
     */
  titre: string;
  /**
     * Description détaillée
     * @maxLength 2000
     */
  description: string;
  statut: TaskStatus;
  createdAt: string;
  updatedAt: string;
}

export interface TaskInput {
  /**
     * Titre de la tâche
     * @minLength 1
     * @maxLength 200
     */
  titre: string;
  /**
     * Description détaillée
     * @minLength 1
     * @maxLength 2000
     */
  description: string;
  statut?: TaskStatus;
}

export interface TaskUpdate {
  /**
     * @minLength 1
     * @maxLength 200
     */
  titre?: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  description?: string;
  statut?: TaskStatus;
}

export interface TaskListResponse {
  success: boolean;
  count: number;
  data: Task[];
}

export interface TaskResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
  data: Task;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  success: boolean;
  message: string;
  errors?: string[];
}