# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API REST du Task Manager — gestion complète des tâches (CRUD)
 * OpenAPI spec version: 0.2.0
 */
import type { TaskStatus } from './taskStatus';

export interface Task {
  /** Identifiant unique */
  id: string;
  /**
     * Titre de la tâche
     * @maxLength 200
     */
  titre: string;
  /**
     * Description détaillée
     * @maxLength 2000
     */
  description: string;
  statut: TaskStatus;
  createdAt: Date;
  updatedAt: Date;
}