# lib/api-zod/src/generated/types/taskListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API REST du Task Manager — gestion complète des tâches (CRUD)
 * OpenAPI spec version: 0.2.0
 */
import type { Task } from './task';

export interface TaskListResponse {
  success: boolean;
  count: number;
  data: Task[];
}