# lib/api-zod/src/generated/types/taskResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API REST du Task Manager — gestion complète des tâches (CRUD)
 * OpenAPI spec version: 0.2.0
 */
import type { Task } from './task';

export interface TaskResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
  data: Task;
}