# lib/api-zod/src/generated/types/taskStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API REST du Task Manager — gestion complète des tâches (CRUD)
 * OpenAPI spec version: 0.2.0
 */

/**
 * Statut possible d'une tâche
 */
export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  À_faire: 'À faire',
  En_cours: 'En cours',
  Terminée: 'Terminée',
} as const;