# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API REST du Task Manager — gestion complète des tâches (CRUD)
 * OpenAPI spec version: 0.2.0
 */
import type { TaskStatus } from './taskStatus';

export interface TaskUpdate {
  /**
     * @minLength 1
     * @maxLength 200
     */
  titre?: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  description?: string;
  statut?: TaskStatus;
}